producer/requirements.txt
google-cloud-pubsub==2.21.5
requests==2.32.3
